#2 [modify]singletonを継承するように変更
